ai-web-studio/requirements.txt
crewai[litellm]>=1.10.0,!=1.10.0
python-dotenv>=1.0.1
